+++ nessai/flows/realnvp.py
@@ -4,6 +4,7 @@ Implementation of Real Non Volume Preserving flows.
 """
 import logging
 
+import numpy as np
 import torch
 import torch.nn.functional as F
 
@@ -32,9 +33,9 @@ class RealNVP(NFlow):
     hidden_features : int
         Number of neurons per layer in each neural network
     num_layers : int
-        Number of coupling tranformations
+        Number of coupling transformations
     num_blocks_per_layer : int
-        Number of layers (or blocks for resnet) per nerual network for
+        Number of layers (or blocks for resnet) per neural network for
         each coupling transform
     mask : array_like, optional
         Custom mask to use between coupling transforms. Can either be
@@ -44,17 +45,17 @@ class RealNVP(NFlow):
     net : {'resnet', 'mlp'}
         Type of neural network to use
     use_volume_preserving : bool, optional (False)
-        Use volume preserving flows which use only additiona and no scaling
+        Use volume preserving flows which use only addition and no scaling
     activation : function
         Activation function implemented in torch
     dropout_probability : float, optional (0.0)
-        Dropout probaiblity used in each layer of the neural network
+        Dropout probability used in each layer of the neural network
     batch_norm_within_layers : bool, optional (False)
        Enable or disable batch norm within the neural network for each coupling
        transform
     batch_norm_between_layers : bool, optional (False)
        Enable or disable batch norm between coupling transforms
-    linear_transform : {'permutaiton', 'lu', 'svd'}
+    linear_transform : {'permutation', 'lu', 'svd', None}
         Linear transform to use between coupling layers. Not recommended when
         using a custom mask.
     """
@@ -84,12 +85,13 @@ class RealNVP(NFlow):
             mask = torch.ones(features)
             mask[::2] = -1
         else:
+            mask = np.array(mask)
             if not mask.shape[-1] == features:
-                raise RuntimeError('Mask does not match number of features')
+                raise ValueError('Mask does not match number of features')
             if mask.ndim == 2 and not mask.shape[0] == num_layers:
-                raise RuntimeError('Mask does not match number of layers')
+                raise ValueError('Mask does not match number of layers')
 
-            mask = torch.from_numpy(mask.astype('float32'))
+            mask = torch.from_numpy(mask).type(torch.get_default_dtype())
 
         if mask.dim() == 1:
             mask_array = torch.empty([num_layers, features])
@@ -114,7 +116,10 @@ class RealNVP(NFlow):
                                          identity_init=True)
                 ])
             else:
-                raise ValueError
+                raise ValueError(
+                    f'Unknown linear transform: {linear_transform}. '
+                    'Choose from: {permutation, lu, svd, None}.'
+                )
 
         if net.lower() == 'resnet':
             from nflows.nn.nets import ResidualNet
@@ -147,7 +152,10 @@ class RealNVP(NFlow):
                         activation=activation)
 
         else:
-            raise RuntimeError(f'Unknown nn type: {net}')
+            raise ValueError(
+                f'Unknown nn type: {net}. '
+                'Choose from: {resnet, mlp}.'
+            )
 
         layers = []
         for i in range(num_layers):